backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==5.0.6
django-cors-headers==4.3.1
django-import-export==4.0.3
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
filelock==3.14.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
mpmath==1.3.0
oauthlib==3.2.2
packaging==24.0
phe==1.5.0
pillow==10.3.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.2.1
pytest-django==4.8.0
python3-openid==3.2.0
requests==2.32.2
requests-file==2.1.0
requests-oauthlib==2.0.0
setuptools==70.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
sympy==1.12
tablib==3.5.0
tldextract==5.1.2
tzdata==2024.1
urllib3==2.2.1
zope.event==5.0
zope.interface==6.4.post2
